Step 3: Add button and a controller
This is where things really started to differ.

First, in Coffeescript, one must be explicit with
`ng-app="pirateBadge"`. Dart didn't care.

Second, let's compare:

Dart:

    bool get inputIsNotEmpty => !name.trim().isEmpty;

Coffeescript:

    $scope.inputIsNotEmpty = -> if $scope.name.match(/\S/) then true else false

Is there a better way to do that in Coffeescript?

No `$scope` in Dart?
--------------------

**In Dart**: the controller declares a selector (`[badges]`), and then
an element in the view is given this selector (`<body badges>`). The
controller is also "published as" something (`ctrl`), and this must be
used to access attributes and functions on the controller
(`{{ctrl.name}}`).

**In Coffeescript**: the controller has a name (`BadgesController`), and
the section of your view that you want to be controlled by it is given a
directive to use that controller (`<body
ng-controller="BadgesController">`). The attributes and functions in the
controller are declared on `$scope`, and are accessed directly in the
view (`{{name}}`).