Step 4: Extract custom component
* AngularDart has a concept of Components, which AngularJS lacks.
  Instead, AngularJS uses a special kind of Directive. Personally, I
  find AngularDart's approach to this much more straightforward.
* The CSS cannot be encapsulated in AngularJS's Directive the way it can
  in AngularDart's component
* AngularDart uses the Shadow DOM, AngularJS does not.

The scoping in AngularJS did not behave at all as expected. Since the
child directive inherits the parent scope, this wasn't clear until I
attempted to move on to step 5. Then I said `<badge
name="{{pirateName()}}` in index.html, but the child wasn't using this
attribute the way I thought it was. Instead, it was using the parent
scope's `name` value. To fix this, I needed the cryptic line `scope: {
name: '@' }`, which isolates the directives scope from that of the
context it's used in. It also sets up a `name` attribute, and the `@`
indicates that this attribute is just passed along to the child
directive as a string. Here again, AngularDart reads much more clearly.